Draft: split Shapestring from Draft.py
